tests: metrics: Improve robustness of PSS mem footprint test
If the docker memory PSS footprint test failed to find the PSS
figures for any of the components (qemu, shim, proxy), then it
would return '0' internally, and still generage a results file,
with incorrect results.
Improve the robustness of the script by trapping any '0' results
where they are not expected.

Reasons smem may fail to find the PSS of a process include:
 - the process name may have changed
 - the process may not be running (may have died for instance)
 - the test is not run with enough priv to read the smaps for the
   process

Fixes: #879

Signed-off-by: Graham Whaley <graham.whaley@intel.com>